% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_compter_obs.R
\docType{data}
\name{data_taxa}
\alias{data_taxa}
\title{Données sur les taxa de la base ASPE}
\description{Tableau regroupant diverses caractéristiques des taxa de la base ASPE}
\usage{data_taxa}
\format{
\code{data_taxa} est \code{data frame} comportant 25 variables :
\describe{
\item{\code{esp_id}}{Identifiants numériques des espèces}
\item{\code{esp_code_alternatif}}{Codes alternatifs des espèces hérités du Conseil Supérieur de la Pêche}
\item{\code{esp_code_sandre}}{Codes sandres des espèces}
\item{\code{esp_code_taxref}}{Codes TAXREF des espèces}
\item{\code{esp_code_idp}}{Codes des espèces utilisés pour l'Indicateur de Dynamique des Population}
\item{\code{esp_nom_commun}}{Noms communs des espèces}
\item{\code{esp_nom_latin}}{Noms latins des espèces}
\item{\code{esp_nom_idp}}{Noms latins des espèces utilisés pour l'Indicateur de Dynamique des Population}
\item{\code{esp_nom_order}}{Noms latins des ordres des espèces}
\item{\code{esp_nom_famille}}{Noms latins des familles des espèces}
\item{\code{esp_hab_tol}}{Tolérances des espèces à la dégradation de leur habitat (\code{intol} : intolérante, \code{tol} : tolérante, \code{med} : intérmediaire)}
\item{\code{esp_qua_tol}}{Tolérances des espèces à la qualité de l'eau (\code{intol} : intolérante, \code{tol} : tolérante, \code{med} : intérmediaire)}
\item{\code{esp_tmp_tol}}{Tolérances des espèces aux variations de température (\code{steno} : sténotherme, \code{eury} : eurytherme)}
\item{\code{esp_oxy_tol}}{Tolérances des espèces à la concentration en oxygène dissous (\code{intol} : intolérante, \code{tol} : tolérante, \code{med} : intérmediaire)}
\item{\code{esp_tox_tol}}{Tolérances des espèces à la pollution (\code{intol} : intolérante, \code{tol} : tolérante, \code{med} : intérmediaire)}
\item{\code{esp_aci_tol}}{Tolérances des espèces à l'acidité (\code{intol} : intolérante, \code{tol} : tolérante, \code{med} : intérmediaire)}
\item{\code{esp_tot_tol}}{Tolérances globales des espèces à la qualité de leur milieu (\code{intol} : intolérante, \code{tol} : tolérante, \code{med} : intérmediaire)}
\item{\code{esp_taille_mat}}{Tailles à 50 \% de maturité des espèces (en mm)}
\item{\code{tlo_id}}{Identifiant du type de taille (1 : longueure à la fouche, 2 : longueure totale)}
\item{\code{a}}{Interceptes de la relation taille-poids des espèces (en g/mm : Poids(g) = a * Taille(mm)^b)}
\item{\code{b}}{Coefficients de la relation taille-poids des espèces (Poids(g) = a * Taille(mm)^b)}
\item{\code{source_ab}}{Source des paramètres des relations taille-poids (estimés à partir des données de la base ASPE ou issus de Fishbase)}
\item{\code{LT_LF}}{Paramètres de la relation entre longueur totale (LT) et la longueur à la fourche (LF) des espèces (LT = LT_LF x LF)}
\item{\code{Native_fr}}{Logique indiquant si les espèces sont indigènes en France métropolitaine}
\item{\code{Naturalise_fr}}{Logique indiquant si les espèces non indigènes sont naturalisées en France métropolitaine}
}
}
\keyword{datasets}
