% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_compter_obs.R
\name{def_compter_obs}
\alias{def_compter_obs}
\title{Définir les attributs des séries d'observations de pêche}
\usage{
def_compter_obs(df, var_id, var_tmp, var_pro = NULL)
}
\arguments{
\item{df}{Dataframe}

\item{var_id}{Nom(s) de la ou des variable(s) contenant les identifiants des points de prélèvement (ex : pop_id, c(sta_id, pop_id))}

\item{var_tmp}{Nom de la variable contenant les dates de prélèvement (ex: annee)}

\item{var_pro}{[optionnel] Nom de la variable contenant les identifiants des protocoles de pêche}
}
\value{
La fonction renvoie un dataframe avec le nombre d'observations (nb_obs), le nombre de dates consécutives sans opération de pêche (nb_na_cons), les dates de début (min_tmp) et de fin (max_tmp) de la série pour chaque point de prélèvement et, si \code{var_pro} est specifié, le nombre de protocoles de pêche (nb_pro), le nombre de changements de protocole (nb_chg) et les protocoles correspondant.
}
\description{
Définie le nombre d'observations, de dates consécutives sans opération de pêche pour chaque points de prélèvement et, si \code{var_pro} est spécifié, le nombre d'observations, de dates consécutives sans opération de pêche, de protocoles de pêche et de changements de protocole pour chaque points de prélèvement et chaque combinaison de protocoles de pêche.
}
\examples{
\dontrun{
## Importation des données
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_ope_date() \%>\%
   mef_ajouter_type_protocole()

## Définir la fréquence et la distribution des observations 
# par points de prélèvement et combinaison de protocoles de pêche 
df_obs <- def_compter_obs(df,
                          var_id = pop_id,
                          var_tmp = annee,
                          var_pro = pro_libelle)
# par points de prélèvement
df_obs <- def_compter_obs(df,
                          var_id = pop_id,
                          var_tmp = annee)
}}
