% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_compter_obs.R
\docType{data}
\name{distribution_native}
\alias{distribution_native}
\title{Distributions natives des taxa}
\description{Fichier cartographique de la distribution native des espèces ou groupes taxonomiques}
\usage{distribution_native}
\format{
\code{distribution_native} est un objet \code{sf} dont le système de projection des coordonnées est le RGF93 Lambert 93 (EPSG : 2154). Cette objet inclus les distributions natives des taxa spécifiés par :
\describe{
\item{\code{esp_code_alternatif}}{Codes alternatifs des espèces hérités du Conseil Supérieur de la Pêche} *
\item{\code{esp_code_idp}}{Codes des espèces utilisés pour l'Indicateur de Dynamique des Population}
\item{\code{esp_code_ref}}{Codes alternatifs des espèces de référence utilisés pour ajouter les espèces nouvellement ou non-décrites}
}
}
\keyword{datasets}
