% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_compter_obs.R
\docType{data}
\name{fond_france}
\alias{fond_france}
\title{Fond de carte de la France métropolitaine}
\description{Fichier cartographique représentant les contours de la France métropolitaine et de la Corse}
\usage{fond_france}
\format{
\code{fond_france} est un objet \code{sf} dont le système de projection des coordonnées est le RGF93 Lambert 93 (EPSG : 2154).
}
\keyword{datasets}
