% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_absence.R
\name{mef_ajouter_absence}
\alias{mef_ajouter_absence}
\title{Ajouter à un dataframe de présences les lignes des absences}
\usage{
mef_ajouter_absence(
  df,
  var_id,
  var_taxon,
  var_abs,
  var_obs = ope_id
)
}
\arguments{
\item{df}{Dataframe de présences}

\item{var_id}{Nom de la ou des variable(s) contenant les identifiants des points de prélèvement (ex : pop_id, sta_id, c(sta_id, saison))}

\item{var_taxon}{Nom de la variable contenant les identifiants des espèces (ex : esp_code_alternatif, esp_code_idp)}

\item{var_abs}{Nom de la ou des variable(s) dont les absences sont à ajouter (ex : lop_effectif, c(lop_effectif, lop_poids)}

\item{var_obs}{[optionnel] Nom de la ou des variable(s) identifiant les observations de pêche (ex : ope_id, c(ope_id, pas_numero), annee)}
}
\value{
La fonction renvoie le dataframe \code{df} avec les lignes supplémentaires des absences (\code{var_abs} = 0).
}
\description{
Ajoute à un dataframe de présences les absences de chaque espèce à partir des informations disponibles pour les autres espèces (l'absence d'une espèce correspond à la non-observation de cette espèce lors d'un évenement de pêche sur un point de prélèvement où elle est recensée au moins une fois).
}
\examples{
\dontrun{
## Ajout des absences sur les effectifs par opérations de pêche (par défaut) 
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_absence(var_id = pop_id,
                       var_taxon = esp_code_alternatif,
                       var_abs = lop_effectif)

## Ajout des absences sur les effectifs et les poids par années
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_poids() \%>\%
   mef_ajouter_ope_date() \%>\%
   summarise(effectif = sum(lop_effectif),
             poids = sum(lop_poids),
             .by=c(pop_id, annee, esp_code_alternatif)) \%>\%
   mef_ajouter_absence(var_id = pop_id,
                       var_taxon = esp_code_alternatif,
                       var_abs = c(effectif,poids),
                       var_obs = annee)
}
}