% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_coords.R
\name{mef_ajouter_coords}
\alias{mef_ajouter_coords}
\title{Ajouter à un dataframe les coordonnées géographiques et le code EPSG}
\usage{
mef_ajouter_coords(df, var_id)
}
\arguments{
\item{df}{Dataframe}

\item{var_id}{Nom de la variable contenant les identifiants des sites de prélèvement (sta_id ou pop_id)}
}
\value{
La fonction renvoie le dataframe \code{df} auquel ont été ajoutés les coordonnées géographiques et le code EPSG associé. 
}
\description{
Ajoute à un dataframe les coordonnées géographiques et le code EPSG associé.
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
mef_ajouter_coords(var_id = pop_id)
}
}
