% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_coords_wgs84.R
\name{mef_ajouter_coords_wgs84}
\alias{mef_ajouter_coords_wgs84}
\title{Ajouter à un dataframe les coordonnées géographiques WGS84}
\usage{
mef_ajouter_coords_wgs84(df, var_id)
}
\arguments{
\item{df}{Dataframe}

\item{var_id}{Nom de la variable contenant les identifiants des sites de prélèvement (sta_id ou pop_id)}
}
\value{
La fonction renvoie le dataframe \code{df} avec les coordonnées géographiques X et Y dans le système de référence WGS84.
}
\description{
Ajoute à un dataframe les coordonnées géographiques dans le système de référence WGS84.
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_coords_wgs84(var_id = pop_id)
}
}