% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_na.R
\name{mef_ajouter_na}
\alias{mef_ajouter_na}
\title{Ajouter à un dataframe de présences ou de présences/absences les observations manquantes}
\usage{
mef_ajouter_na(
  df,
  var_id,
  var_taxon,
  var_obs,
  vec_obs = NULL
)
}
\arguments{
\item{df}{Dataframe}

\item{var_id}{Nom de la ou des variable(s) contenant les identifiants des points de prélèvement (ex : pop_id, sta_id, c(sta_id, saison)}

\item{var_taxon}{Nom de la variable contenant les identifiants des espèces (ex : esp_code_alternatif, esp_code_idp)}

\item{var_obs}{Nom de la ou des variable(s) identifiant les observations de pêche (ex : ope_id, annee, c(annee, stade)}

\item{vec_obs}{[optionnel] Vecteur listant l'ensemble des observations de pêche attendues (par défaut, correspond à l'ensemble des \code{var_obs} du dataframe \code{df})}
}
\value{
La fonction renvoie le dataframe \code{df} avec les lignes supplémentaires des valeurs manquantes.
}
\description{
Ajoute à un dataframe de présences ou ou de présences/absences les observations manquantes de chaque espèce (i.e. absences d'évenements de pêche sur les points de prélèvement où l'espèce est recensée).
}
\examples{
\dontrun{
## Importation et sélection des données
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_ope_date() \%>\%
   filter(annee \%in\% 2010:2020, !annee == 2015)

## Ajout des années manquantes à partir des années du dataframe df 
df_na <- mef_ajouter_na(df,
                        var_id = pop_id,
                        var_taxon =esp_code_alternatif,
                        var_obs = annee)
sort(unique(df_na$annee))
# dans ce cas : vec_obs = c(2010:2014,2016:2020) 
# car l'année 2015 n'est pas présente dans le dataframe df

## Ajout des années manquantes à partir d'un vecteur (inclusion de l'année 2015)
df_na <- mef_ajouter_na(df,
                        var_id = pop_id,
                        var_taxon =esp_code_alternatif,
                        var_obs = annee,
                        vec_obs = 2010:2020)
sort(unique(df_na$annee))
}
}
