% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_ope_date_complete.R
\name{mef_ajouter_ope_date_complete}
\alias{mef_ajouter_ope_date_complete}
\title{Ajouter à la passerelle la date complète de chaque opération}
\usage{
mef_ajouter_ope_date_complete(df)
}
\arguments{
\item{df}{Dataframe devant contenir la variable "ope_id"}
}
\value{
La fonction renvoie le data frame auquel à été ajouté la date (ope_date), l'année, le mois et le jour de chaque opération de pêche
}
\description{
Ajoute à un dataframe la date, l'année, le mois et le jour de chaque opération de pêche
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_ope_date_complete()
}
}