% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_ope_saison.R
\name{mef_ajouter_ope_saison}
\alias{mef_ajouter_ope_saison}
\title{Ajouter à la passerelle la saison de chaque opération}
\usage{
mef_ajouter_ope_saison(df)
}
\arguments{
\item{df}{Dataframe devant contenir la variable "ope_date" ou la variable "mois"}
}
\value{
La fonction renvoie le dataframe \code{df} auquel à été ajouté la saison de chaque opération de pêche.
}
\description{
Ajoute à un dataframe la saison de chaque opération de pêche, i.e. printemps (avril à juillet) ou automne (août à novembre). 
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_ope_date() \%>\%
   mef_ajouter_ope_saison()
}
}