% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_poids.R
\name{mef_ajouter_poids}
\alias{mef_ajouter_poids}
\title{Ajouter les poids des lots à la passerelle}
\usage{
mef_ajouter_poids(df)
}
\arguments{
\item{df}{Dataframe devant contenir la variable "lop_id"}
}
\value{
La fonction renvoie le dataframe \code{df} auquel ont été ajoutés les poids des lots ("lop_poids" et "lop_poids_estime")
}
\description{
Ajoute à un dataframe les poids des lots.
}
\examples{
\dontrun{
passerelle <- mef_creer_passerelle() \%>\%
           mef_ajouter_poids()
}
}
