% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_taille_min_max.R
\name{mef_ajouter_taille_min_max}
\alias{mef_ajouter_taille_min_max}
\title{Ajouter les tailles minimales et maximales des lots}
\usage{
mef_ajouter_taille_min_max(df)
}
\arguments{
\item{df}{Dataframe devant contenir les variables "lop_id" et "tyl_id"}
}
\value{
La fonction renvoie le dataframe \code{df} auquel ont été ajoutés les tailles minimales ("lop_longueur_specimens_taille_mini") et maximales ("lop_longueur_specimens_taille_maxi") des lots.
} 
\description{
Ajoute les tailles minimales et maximales des lots à partir des mesures individuelles ou des longueurs minimales et maximales de la table "lot_poissons" lorsque les individus ne sont pas individuellement mesurés.
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_taille_min_max()
}
}
