% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_ajouter_trait.R
\name{mef_ajouter_trait}
\alias{mef_ajouter_trait}
\title{Ajouter à un dataframe les traits des espèces}
\usage{
mef_ajouter_trait(
  df,
  var_taxon,
  trait
)
}
\arguments{
\item{df}{Dataframe}

\item{var_taxon}{Nom de la variable contenant les identifiants des espèces ("esp_id", "esp_code_alternatif", "esp_code_sandre", "esp_code_taxref" ou "esp_code_idp")}

\item{trait}{[optionnel] Vecteur de caractères listant les traits à ajouter (par défaut, la fonction ajoute les variables "Native_fr", "Naturalise_fr" et "esp_tot_tol")}
}
\value{
La fonction renvoie le dataframe \code{df} auquel ont été ajoutés les traits demandés. 
}
\seealso{
\code{\link{data_taxa}} pour obtenir la liste des traits disponibles.
}
\description{
Ajoute à un dataframe les traits des espèces à partir du dataframe \code{data_taxa}.
}
\examples{
\dontrun{
## Ajout des traits par défaut
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_trait(var_taxon = esp_code_alternatif)

## Ajout des traits liés aux diverses tolérances des espèces
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_trait(var_taxon = esp_code_alternatif,
                     trait =c("esp_hab_tol","esp_qua_tol","esp_tmp_tol",
                              "esp_oxy_tol","esp_tox_tol","esp_aci_tol"))
}
}
