% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_convertir_en_taille_totale.R
\name{mef_convertir_en_taille_totale}
\alias{mef_convertir_en_taille_totale}
\title{Convertir les longueurs à la fourche en longueurs totale}
\usage{
mef_convertir_en_taille_totale(
  df,
  var_tai = mei_taille,
  var_taxon = esp_code_alternatif
)
}
\arguments{
\item{df}{Dataframe devant contenir la variable "tlo_id"}

\item{var_tai}{[optionnel] Nom(s) de la ou des variable(s) contentant les longueurs à convertir}

\item{var_taxon}{[optionnel] Nom de la variable contentant les identifiants des espèces}
}
\value{
La fonction renvoie le dataframe \code{df} après conversion des longueurs à la fourche ("tlo_id" = 1) en longueurs totales ("tlo_id" = 2) et modification de la variable "tlo_id".
}
\description{
Convertie les longueurs à la fourche en longueurs totale à partir des paramètres de conversion de la table \code{\link{data_taxa}}.
}
\examples{
\dontrun{
## Convertir les tailles individuelles
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_mei() \%>\%
   mef_convertir_en_taille_totale()

## Convertir les tailles minimales et maximales des lots
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_taille_min_max() \%>\%
   mef_convertir_en_taille_totale(var_tai = c(lop_longueur_specimens_taille_mini,
                                              lop_longueur_specimens_taille_maxi))
}
}
