% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_estimer_poids_taille.R
\name{mef_estimer_poids_taille}
\alias{mef_estimer_poids_taille}
\title{Estimer le poids des lots à partir de la relation taille-poids}
\usage{
mef_estimer_poids_taille(df, var_taxon = esp_code_alternatif)
}
\arguments{
\item{df}{Dataframe devant contenir les variables "lop_id", "tlo_id", "tyl_id" et "lop_effectif"}

\item{var_taxon}{[optionnel] Nom de la variable contentant les identifiants des espèces}
}
\value{
La fonction renvoie le dataframe \code{df} auquel ont été ajoutés les poids des lots estimés à partir de la relation taille-poids ("lop_poids_taille").
}
\description{
Estime le poids des lots à partir des mesures individuelles et des paramètres de la relation taille-poids de la table \code{\link{data_taxa}}. Pour les lots dont la taille individuelle n'est mesurée que pour un sous-échantillon représentatif, le poids est estimé après reconstitution des effectifs par tailles.
}
\examples{
\dontrun{
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_estimer_poids_taille()
}
}
