% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_distribution_native.R
\name{mef_filtrer_distribution_native}
\alias{mef_filtrer_distribution_native}
\title{Filtrer un dataframe en fonction de la distribution native des espèces}
\usage{
mef_filtrer_distribution_native(
  df,
  var_taxon,
  var_obs = NULL,
  recode_taxon = FALSE
)
}
\arguments{
\item{df}{Dataframe devant contenir les variables X et Y dans le système de projection WGS84}

\item{var_taxon}{Nom de la variable contenant les identifiants des espèces ("esp_code_alternatif" ou "esp_code_idp")}

\item{var_obs}{[optionnel] Nom de la ou des variable(s) contenant les identifiants des observations (utilisé si \code{recode_taxon} = \code{TRUE} : permet de prendre en compte la présence conjointe d'une espèce nouvellement décrite à ajouter et de son espèce de référence lors d'une même observation)}

\item{recode_taxon}{[optionnel] Logique indiquant si les espèces nouvellement ou non-décrites doivent être ajoutées (par défaut, ces espèces ne sont pas ajoutées)}
}
\value{
La fonction renvoie le dataframe \code{df} après suppression, pour chaque espèce, des points de prélèvement non-inclus dans leur distribution native. Si \code{recode_taxon} = \code{TRUE}, les espèces nouvellement ou non-décrites dans la base ASPE sont ajoutées au dataframe.
}
\seealso{
\code{\link{distribution_native}} pour obtenir la liste des espèces dont la distribution native est disponible.
}
\description{
Filtre un dataframe en sélectionnant, pour chaque espèce, les points de prélèvement inclus dans leur distribution native. Permet de prendre en compte les espèces nouvellement ou non-décrites dans la base ASPE à partir d'espèces de référence. 
}
\examples{
\dontrun{
## Filtrer le dataframe sans ajout des espèces non-décrites dans ASPE
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_coords_wgs84(var_id = pop_id) \%>\%
   mef_filtrer_distribution_native(var_taxon = esp_code_alternatif)

## Filtrer le dataframe en ajoutant les espèces nouvellement ou non-décrites dans ASPE   
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_coords_wgs84(var_id = pop_id) \%>\%
   mef_filtrer_distribution_native(var_taxon = esp_code_alternatif,
                                   recode_taxon = TRUE)
# prise en compte lors d'une même opération de pêche de la possible co-occurrence 
# des espèces nouvellement décrites et de leur(s) espèce(s) de référence                                     
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_coords_wgs84(var_id = pop_id) \%>\%
   mef_filtrer_distribution_native(var_taxon = esp_code_alternatif,
                                   var_obs = ope_id,
                                   recode_taxon = TRUE)
}
}
