% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_obs.R
\name{mef_filtrer_obs}
\alias{mef_filtrer_obs}
\title{Filtrer les séries d'observations de pêche}
\usage{
mef_filtrer_obs(
  df,
  df_obs,
  var_id,
  var_id_sup = NULL,
  var_pro = NULL,
  min_obs,
  max_na_cons = NULL,
  max_pro = NULL,
  max_chg = NULL
)
}
\arguments{
\item{df}{Dataframe contenant les observations}

\item{df_obs}{Dataframe contenant les attributs des séries (tel que retourné par la fonction \code{def_compter_obs})}

\item{var_id}{Nom(s) de la ou des variable(s) contenant les identifiants des points de prélèvements (ex : pop_id, c(sta_id, pop_id), c(pop_id, saison))}

\item{var_id_sup}{[optionnel] Nom(s) de la ou des variable(s) contenant les identifiant additionnels des points de prélèvements (ex : pop_id, saison, c(pop_id, saison))}

\item{var_pro}{[optionnel] Nom de la variable contenant les identifiants des protocoles de pêche}

\item{min_obs}{Numerique définissant le nombre minimal d'observation par points de prélèvement (\code{var_id})}

\item{max_na_cons}{Numerique définissant le nombre maximal de dates consécutives sans opération de pêche par points de prélèvement (\code{var_id})}

\item{max_pro}{Numerique définissant le nombre maximal de protocoles de pêche par points de prélèvement (\code{var_id} et \code{var_id_sup})}

\item{max_chg}{Numerique définissant le nombre maximal de changement de protocole de pêche par points de prélèvement (\code{var_id} et \code{var_id_sup})}
}
\value{
La fonction renvoie le dataframe \code{df} après suppression des points de prélèvement et/ou des dates de pêche ne correspondant pas aux critères specifiés.

Lorsque plusieurs séries d'observations définis par \code{var_id} répondent aux critères, la fonction conserve la série comportant le nombre maximal d'observations et le nombre minimal de dates consécutives sans pêche.
}
\seealso{
\code{\link{def_compter_obs}}
}
\description{
Filtre un dataframe \code{df} contenant les séries d'observations de pêche à partir d'un dataframe \code{df_obs} contenant les attributs des séries et de critères sur la fréquence et la distribution des observations.
}
\examples{
\dontrun{
## Importation et sélection des données
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_ope_date() \%>\%
   mef_ajouter_ope_saison() \%>\%
   mef_ajouter_type_protocole() \%>\%
   filter(!is.na(saison)) %>%
   filter(pop_id \%in\% c(10,2271,3969,7405,12316,35119,71350,79638,86743)) \%>\%
   filter(annee \%in\% 2010:2020)

## Définir la fréquence et la distribution des observations 
# par points de prélèvement, saison et combinaison de protocoles de pêche 
df_obs <- def_compter_obs(df,
                          var_id=c(pop_id,saison),
                          var_tmp=annee,
                          var_pro=pro_libelle)
                                                    
## Filter df en fonction de df_obs
# conservation d'une seule serie d'observation par points de prélèvement                                     
df_filtre <- mef_filtrer_obs(df,
                             df_obs,
                             var_id=pop_id,
                             var_id_sup=saison,
                             var_pro=pro_libelle,
                             min_obs = 7,
                             max_na_cons = 3,
                             max_pro = 2,
                             max_chg = 1)
# conservation d'une seule serie d'observation par points de prélèvement et saison  
df_filtre <- mef_filtrer_obs(df,
                             df_obs,
                             var_id=c(pop_id,saison),
                             var_pro=pro_libelle,
                             min_obs = 7,
                             max_na_cons = 3,
                             max_pro = 2,
                             max_chg = 1)
# par points de prélèvement sans filtre sur les protocoles
df_filtre <- mef_filtrer_obs(df,
                             df_obs,
                             var_id=pop_id,
                             min_obs = 7,
                             max_na_cons = 3)
}
}