% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_operation.R
\name{mef_filtrer_operation}
\alias{mef_filtrer_operation}
\title{Filtrer les opérations de pêche}
\usage{
mef_filtrer_operation(
  df,
  var_id,
  var_tmp,
  var_surf = NULL,
  var_pro = NULL,
  var_pds = NULL,
  var_mei = NULL,
  var_pas = NULL,
  var_date = NULL,
  default = FALSE
)
}
\arguments{
\item{df}{Dataframe devant contenir le champ "ope_id"}

\item{var_id}{Nom(s) de la ou des variable(s) contenant les identifiants des points de prélèvement (ex : sta_id, pop_id, c(pop_id, saison)}

\item{var_tmp}{Nom de la variable contenant les dates de prélèvement (ex: annee)}

\item{var_surf}{[optionnel] Nom de la variable contenant les surfaces de pêche}

\item{var_pro}{[optionnel] Nom de la variable contenant les protocoles de pêche}

\item{var_pds}{[optionnel] Nom de la variable contenant les poids}

\item{var_mei}{[optionnel] Nom de la variable contenant des identifiants sur les mesures individuelles (ex : mei_id)}

\item{var_pas}{[optionnel] Nom de la variable contenant les numéros de passage de pêche}

\item{var_date}{[optionnel] Nom de la variable contenant les dates additionnelles de pêche (ex : mois, saison)}

\item{default}{Logique : si TRUE, selection de la première opération de pêche)}
}
\value{
La fonction renvoie le dataframe \code{df} après suppression des opérations de pêche en doublon.  
}
\description{
Lorsque plusieurs opérations de pêche par points de prélèvement et par date de prélèvement ont eu lieu, filtre les opérations selon différents critères optionnels : (1) sélection préférencielle des opérations dont la surface de pêche est renseignée, (2) sélection des opérations minimisant le nombre de changements de protocoles de pêche sur l'ensemble de la série temporelle définie par points de prélèvement, (3) sélection préférencielle des opérations dont le poids est renseignée, (4) sélection préférencielle des opérations associées à des mesures individuelles, (5) sélection préférencielle des opérations associées à plusieurs passages de pêche, (6) sélection des opérations fournissant la série la plus homogène en terme de date.
}
\examples{
\dontrun{
## Conservation d'une opération de pêche par points de prélèvement et par an
# en fonction de la surface, puis du protocole de pêche
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_ope_date_complete() \%>\%
   mef_ajouter_surf_calc() \%>\%
   mef_ajouter_type_protocole() \%>\%
   mef_filtrer_operation(var_id = pop_id,
                         var_tmp = annee,
                         var_surf = ope_surface_calculee,
                         var_pro = pro_libelle,
                         default = TRUE)
# en fonction du mois de pêche, puis de la surface et du protocole de pêche                         
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_ope_date_complete() \%>\%
   mef_filtrer_operation(var_id = pop_id,
                         var_tmp = annee,
                         var_date = mois) \%>\%
   mef_ajouter_surf_calc() \%>\%
   mef_ajouter_type_protocole() \%>\%
   mef_filtrer_operation(var_id = pop_id,
                         var_tmp = annee,
                         var_surf = ope_surface_calculee,
                         var_pro = pro_libelle,
                         default = TRUE)
}
}
