% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_filtrer_presence.R
\name{mef_filtrer_presence}
\alias{mef_filtrer_presence}
\title{Filtrer un dataframe en fonction de la détection des espèces}
\usage{
mef_filtrer_presence(df, var_id, var_taxon, var_obs)
}
\arguments{
\item{df}{Dataframe}

\item{var_id}{Nom de la ou des variable(s) contenant les identifiant des points prélèvement (ex : sta_id, pop_id, c(sta_id, pop_id))}

\item{var_taxon}{Nom de la variable contenant les identifiants des espèces (ex : esp_code_alternatif, esp_code_sandre, esp_code_idp)}

\item{var_obs}{Nom de la variable contenant les présences/absence (numérique ; ex : lop_effectif, lop_poids)}
}
\value{
La fonction renvoie le dataframe \code{df} après suppression, pour chaque espèce, des points de prélèvement ne contenant que des valeurs nulles ou manquantes (\code{var_obs} = c(0,NA)). 
}
\description{
Filtre un dataframe en sélectionnant, pour chaque espèce, les points de prélèvements où l'espèce est détectée au moins une fois. 
}
\examples{
\dontrun{
## Création d'un dataframe de présences/absences
df <- mef_creer_passerelle() \%>\%
   mef_ajouter_lots() \%>\%
   mef_ajouter_passage() \%>\%
   summarise(effectif = sum(lop_effectif),
             .by = c(pop_id,ope_id,pas_numero,esp_code_alternatif)) \%>\%
    mef_ajouter_absence(var_id = pop_id,
                        var_taxon = esp_code_alternatif,
                        var_abs = effectif,
                        var_obs= c(ope_id,pas_numero))
                        
## Filtrer le dataframe en fonction de la présences des espèces lors du premier passage de pêhce
df_filtre <- df \%>\%
          filter(pas_numero \%in\% 1) \%>\%
          mef_supprimer_site_absence(var_id = pop_id, 
                                     var_taxon = esp_code_alternatif,
                                     var_obs = effectif)
}
}
