% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mef_reconstituer_capture_par_stade.R
\name{mef_reconstituer_capture_par_stade}
\alias{mef_reconstituer_capture_par_stade}
\title{Reconstituer les effectifs et les poids par espèce et stade de vie}
\usage{
mef_reconstituer_capture_par_stade(
  df,
  var_id = ope_id,
  var_taxon = esp_code_alternatif,
  var_tai = esp_taille_mat,
  var_tlo = tlo_id,
  tb = NULL
)
}
\arguments{
\item{df}{Dataframe devant contenir les variables "tlo_id", "tai_effectif", "tai_poids", "tai_poids_estime", "lop_effectif","lop_poids", "lop_poids_estime", "mei_taille", "lop_longueur_specimens_taille_mini" et "lop_longueur_specimens_taille_maxi"}

\item{var_id}{[optionnel] Nom(s) de la ou des variable(s) contenant les identifiants des observations (ex : lop_id, ope_id, c(pop_id, ope_id))}

\item{var_taxon}{[optionnel] Nom de la variable contentant les identifiants des espèces}

\item{var_tai}{[optionnel] Nom de la variable identifiant les tailles de reférence utilisées pour définir les stades (par défaut, les tailles à maturité du dataframe \code{data_taxa})}

\item{var_tlo}{[optionnel] Nom de la variable identifiant le type de longueur}

\item{tb}{[optionnel] Dataframe fournissant les tailles de référence utilisées pour définir les stades}
}
\value{
La fonction renvoie un dataframe avec les effectifs ("std_effectif"), les poids mesurés ("std_poids"), les poids estimés ("std_poids_estime") et les poids calculé via la relation taille-poids ("std_poids_taille") par observation, espèce et stade.
}
\seealso{
\code{\link{mef_reconstituer_capture_par_taille}} et \code{\link{mef_reconstituer_effectif_par_stade}}.
}
\description{
Reconstitue les effectifs et les poids par espèce et selon deux stades de vie ("S1" lorsque la taille des individus est strictement inférieure à la taille de référence de l'espèce ou "S2"). 
}
\examples{
 \dontrun{
 df <- mef_creer_passerelle() \%>\%
    mef_ajouter_lots() \%>\%
    mef_ajouter_poids() \%>\%
    mef_reconstituer_capture_par_taille() \%>\%
    mef_reconstituer_capture_par_stade()
 }
}
