% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_repartition_observation.R
\name{plot_repartition_observation}
\alias{plot_repartition_observation}
\title{Cartes de répartition des observations de chaque espèce}
\usage{
plot_repartition_observation(
  df,
  var_taxon,
  distrib_native = TRUE,
  titre = TRUE,
  var_titre = NULL,
  cex.titre = 2,
  font.titre = 2,
  cex.point = 0.75,
  col.point = "black",
  col.distribution = "grey90",
  col.secteur = "grey",
  col.contour = "black",
  new.plot = TRUE,
  sauv.png = NULL,
  sauv.eps = NULL
)
}
\arguments{
\item{df}{Dataframe devant contenir les variables X et Y dans le système de projection WGS84}

\item{var_taxon}{Nom de la variable contenant les identifiants des espèces ("esp_code_alternatif" ou "esp_code_idp" si \code{distrib_native = TRUE})}

\item{distrib_native}{[optionnel] Logique indiquant si la distribution native des espèces est affichée}

\item{titre}{[optionnel] Logique indiquant si un titre est affiché}

\item{var_titre}{[optionnel] Nom de la variable utilisée pour le titre si différente de \code{var_taxon} (ex : esp_nom_latin, esp_nom_commun ; cf. \code{\link{data_taxa}} pour obtenir la liste des variables disponibles)}

\item{cex.titre}{[optionnel] Numérique définissant la taille du titre}

\item{font.titre}{[optionnel] Numérique définissant la police du titre (1: titre normale, 2: titre en gras, 3: titre en italique, 4: titre en italique gras)}

\item{cex.point}{[optionnel] Numérique définissant la taille des points représentant les observations}

\item{col.point}{[optionnel] Caractère définissant la couleur des points représentant les observations}

\item{col.distribution}{[optionnel] Caractère définissant la couleur utilisée pour représenter la distribution native (uniquement si \code{distrib_native = TRUE})}

\item{col.secteur}{[optionnel] Caractère définissant la couleur des contours des secteurs hydrographiques}

\item{col.contour}{[optionnel] Caractère définissant la couleur du contour de la carte}

\item{new.plot}{[optionnel] Logique indiquant si une nouvelle fênetre graphique est générée pour chaque espèce}

\item{sauv.png}{[optionnel] Caractère définissant le chemin et le nom des fichiers pour un enregistrement des cartes en ".png"}

\item{sauv.eps}{[optionnel] Caractère définissant le chemin et le nom des fichiers pour un enregistrement des cartes en ".eps"}
}
\value{
La fonction renvoie les cartes de distribution des observations de chaque espèce. 
}
\description{
Génère les cartes donnant pour chaque espèce la répartition des observations.
}
\examples{
\dontrun{
## Importation et sélection des données  
df <- mef_creer_passerelle() \%>\%
  mef_ajouter_ope_date() \%>\%
  filter(annee \%in\% 2010:2020) \%>\%
  mef_ajouter_lots() \%>\%
  mef_ajouter_coords_wgs84(var_id = pop_id) \%>\%
  filter(esp_code_alternatif \%in\% c("BRO","CHA"))

## Cartes de répartition des observations des espèces 
plot_repartition_observation(df, 
                              var_taxon = esp_code_alternatif)

# supprimer l'affichage de la distribution native
plot_repartition_observation(df,
                              var_taxon = esp_code_alternatif,
                              distrib_native = FALSE)
                              
# modifier le titre (nom latins des espèces en italique gras) 
# et sauvegarder dans le répertoire courant au format .png
plot_repartition_observation(df, 
                              var_taxon = esp_code_alternatif,
                              var_titre = esp_nom_latin,
                              font.titre=4,
                              sauv.png = "Carte")

# modifier la couleur utilisée pour les distributions natives
# et supprimer l'affichage des secteurs hydrographiques                              
plot_repartition_observation(df,
                              var_taxon = esp_code_alternatif,
                              var_titre = esp_nom_latin,
                              font.titre=4,
                              col.distribution="azure",
                              col.secteur="transparent",
                              sauv.png = "Carte")

# supprimer le titre                              
plot_repartition_observation(df,
                              var_taxon = esp_code_alternatif,
                              titre = FALSE,
                              col.distribution="azure",
                              col.secteur="transparent",
                              sauv.png = "Carte")

# afficher les cartes dans la même fenêtre graphique
par(mfrow=c(1,2))
plot_repartition_observation(df, 
                              var_taxon = esp_code_alternatif,
                              new.plot=FALSE)
}
}